Don't emit event to update global template options.
